requirements.txt
Flask
Flask-SocketIO
eventlet
SQLAlchemy
Flask-SQLAlchemy
python-socketio[client]
websocket-client
Pillow
numpy